+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Grid
     * [Added argument initialRecord to grid columns render function](https://github.com/softindex/uikernel/pull/249)
     * [Added grid context param to columns render method](https://github.com/softindex/uikernel/pull/269)
+    * [Added `multipart/form-data` encoding to GridXHRModel](https://github.com/softindex/uikernel/pull/270)
     * Added `recordId` property to GridModel.prototype.isValidRecord
     
   * Form